+++ src/components/tree_view/index.ts
@@ -1,0 +1,1 @@
+export { EuiTreeView } from './tree_view';